Fixed for user input to be searched.
The original program only allowed for one site to be searched and it had to be imbedded in the code before use, was not truly done. Now user input will be used for scrapping sites but now there is only "Ingredients and Instructions" that print out to UI. Need to fix for description, yield and other info.